Handle long messages in MQL SHA implementations

+++ sha256.hpp
@@ -72,7 +72,7 @@ namespace hmac_hash {
 
     protected:
         void transform(const uint8_t *message, size_t block_nb);
-        size_t m_tot_len;
+        uint64_t m_tot_len;
         size_t m_len;
         uint8_t m_block[2 * SHA224_256_BLOCK_SIZE];
         uint32_t m_h[8];
